requirements.txt
APScheduler==3.0.0
requests==2.22.0
python-dotenv==0.10.3
postmarker==0.15.0